[fix] See below the list of important fixes and updates:
 - in sequence_matching.py
Checking for empty/junk text patterns
is_empty = any(token for token in pattern if token not in junk) -->  correct = any([token not in junk for token in pattern])
if not is_empty --> if correct

 - added LIMIT in ml_clusterization.py
The reason of this limitation is possible cases when a number of clusters is too big and executing of LogCluster (perl-based) app to search common patterns for each cluster may be too long. So, now we use LogCluster only if the number of clusters is less than the limit (which is now set to 30).

 - pyTextRank was temporary removed (I left it just for tests)

 - requirements.txt was updated with newer versions of libraries